Add in the new filter api
This introduces a trait: Filterable that is implemented for various types.

This'll produce a Filter type which can be used to
filter the tree from a starting node and with a
predicate.

This adds in a `Properties` projection of a LayoutNode

Also, Rasterizer::blend is more flexible
Also, new compact_tree mode is available