New style system, take 3.
As discussed in #42 this associates the style with the view.

This implements the idea by introducing a Style<Arg = T> trait that styles implement.

The arg for the style is the arg passed to the deferred 'class' method.

StyleKind has been hidden and a more usable/reusable ApplicableStyle<S: Style> type has been added.

Builder now takes a Style associated type and has 2 default methods:
- `class(self, impl Fn(&Palette, Self::Style::Args) -> Self::Style) -> Self `
- `style(self, Self::Style) -> Self`

With these changes `Builder` and `Style` will likely a required import for users defining per-view styling.