New style system (#48)
* Add basic Painter implementation

This also changes `TextShape` to store a `Str`
rather than a `Cow<'a, str>`.

Add a few examples

* add a color_select example

* Properties is now public

* remove old references to ui.adhoc now that Adhoc is gone

* Add Style/Class to the Builder trait

This is part of #42

This adds two associated types, Class and Style to the builder

and provides a default no-op class/style builder
method that takes in each of these types and
returns the builder.

* bump deps

* Update documentation for Builder styling

* fix some clippy lints

* New style system, take 3.

As discussed in #42 this associates the style with the view.

This implements the idea by introducing a Style<Arg = T> trait that styles implement.

The arg for the style is the arg passed to the deferred 'class' method.

StyleKind has been hidden and a more usable/reusable ApplicableStyle<S: Style> type has been added.

Builder now takes a Style associated type and has 2 default methods:
- `class(self, impl Fn(&Palette, Self::Style::Args) -> Self::Style) -> Self `
- `style(self, Self::Style) -> Self`

With these changes `Builder` and `Style` will likely a required import for users defining per-view styling.