Improve key input dispatching.
This renames `Interest::FOCUS_INPUT` to `Interest::INPUT`.

This also changes the logic of how `InputState::update_key_event()` works.

If a view is focused, it'll only dispatch to it.
Otherwise it walks the trees dispatching events
until a `Handle::Sink` is returned.

This closes #33.

A focus-ring/tab-indices was not implemented. That'll be a future addition.