Yet another attempt at a nice styling api.
This one is a lot better and unifies a lot of ideas.

Now the Builder only needs to optionally return the views ApplicableStyle.

With this, the ViewExt can do all sorts of things like:
- style
- style_if
- class
- class_if
- or
- disabled
- hoverable
- disabled_if
- hoverable_if

This introduces a 'StyleState' which can be used to determine whether a view is:
- hoverable
- disabled

To apply a style, `style.apply()` now takes a `StyleApplicator` (a `&Render` or `&Ui`) which returns the evaluated style

Instead of weird `args: Self::Args` in the styles, we now have a `StyleOptions<Self::Args>` which has:
- hovered: bool
- focused: bool
- selected: bool
- interactive: bool
- state: StyleState
- args: Self::Args

`style.apply()` now also takes a closure for setting the 'args' (the current state from the View for the style to use)